LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssSkewX.tentative-expected.txt


FAIL Constructing a CSSSkewX with a keyword throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a double throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a unitless zero throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a string angle throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a number CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a time dimension CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSSkewX with a CSSMathValue of length type throws a TypeError assert_throws_js: function "() => new CSSSkewX(value)" threw object "ReferenceError: Can't find variable: CSSSkewX" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSSkewX.ax with a keyword throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a double throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a unitless zero throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a string angle throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a number CSSUnitValue throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a time dimension CSSUnitValue throws a TypeError Can't find variable: CSSSkewX
FAIL Updating CSSSkewX.ax with a CSSMathValue of length type throws a TypeError Can't find variable: CSSSkewX
FAIL CSSSkewX can be constructed from an angle CSSUnitValue Can't find variable: CSSSkewX
FAIL CSSSkewX can be constructed from a CSSMathValue of angle type Can't find variable: CSSSkewX
FAIL CSSSkew.ax can be updated to an angle CSSUnitValue Can't find variable: CSSSkewX
FAIL CSSSkew.ax can be updated to a CSSMathValue of angle type Can't find variable: CSSSkewX
FAIL Modifying skewX.is2D is a no-op Can't find variable: CSSSkewX
